requirements.txt
numpy
PyYAML
requests
kafka-python
